refactor: move `SignalEvent` into `strategy/mod.rs` & update imports accordingly
