+++ src/event.rs
@@ -1,9 +1,9 @@
 use crate::data::MarketEvent;
+use crate::strategy::{SignalEvent, SignalForceExit};
 use crate::execution::FillEvent;
 use crate::portfolio::OrderEvent;
 use crate::portfolio::position::{EquityPoint, Position, PositionExit, PositionUpdate};
 use crate::portfolio::repository::{AvailableCash, TotalEquity};
-use crate::strategy::signal::{SignalEvent, SignalForceExit};
 use chrono::{DateTime, Utc};
 use serde::Serialize;
 use std::fmt::Debug;
